CSS
I changed the position of the car.
I aligned the titles and text.
I added a contact button.
I added indicators, such as the car number and also 3 horizontal lines that will change color depending on the selected car, indicating whether the "page" is in the first, second or third car.